chore: making it compatible with js
