+++ gradle.properties
@@ -12,3 +12,5 @@ org.gradle.daemon=false
 kotlin.code.style=official
 # Memory Optimisation
 org.gradle.jvmargs=-Xmx4096M
+# Disable these warnings because using api propagates libraries and we don't want that on commonMain
+kotlin.suppressGradlePluginWarnings=IncorrectCompileOnlyDependencyWarning
