+++ packages/mongodb-access-adapter/src/jvmMain/kotlin/com/mongodb/jbplugin/accessadapter/slice/GetCollectionSchema.jvm.kt
@@ -1,0 +1,39 @@
+package com.mongodb.jbplugin.accessadapter.slice
+
+import com.mongodb.jbplugin.mql.BsonAnyOf
+import com.mongodb.jbplugin.mql.BsonArray
+import com.mongodb.jbplugin.mql.BsonNull
+import com.mongodb.jbplugin.mql.BsonObject
+import com.mongodb.jbplugin.mql.BsonType
+import com.mongodb.jbplugin.mql.primitiveOrWrapper
+import com.mongodb.jbplugin.mql.toBsonType
+
+actual fun recursivelyBuildSchema(value: Any?): BsonType {
+    return when (value) {
+        null -> BsonNull
+        is Map<*, *> -> BsonObject(
+            value.map {
+                it.key.toString() to
+                    recursivelyBuildSchema(it.value)
+            }.toMap()
+        )
+
+        is Collection<*> -> recursivelyBuildSchema(value.toTypedArray())
+        is Array<*> ->
+            BsonArray(
+                value
+                    .map {
+                        it?.javaClass?.toBsonType(it) ?: BsonNull
+                    }.toSet()
+                    .let {
+                        if (it.size == 1) {
+                            it.first()
+                        } else {
+                            BsonAnyOf(it)
+                        }
+                    },
+            )
+
+        else -> primitiveOrWrapper(value.javaClass).toBsonType()
+    }
+}
