+++ packages/mongodb-mql-engines/src/commonMain/kotlin/com/mongodb/jbplugin/indexing/CollectionIndexConsolidation.kt
