+++ packages/mongodb-mql-engines/src/commonMain/kotlin/com/mongodb/jbplugin/indexing/IndexAnalyzer.kt
