+++ packages/mongodb-mql-engines/src/commonMain/kotlin/com/mongodb/jbplugin/linting/IndexCheckingLinter.kt
